+++ plugin-animation/src/main/java/com/mapbox/maps/plugin/animation/animator/CameraAnimator.kt
@@ -49,6 +49,7 @@ abstract class CameraAnimator<out T> (
 
   internal var canceled = false
   internal var isInternal = false
+  internal var skipped = false
 
   init {
     setObjectValues(targets[0], targets[0])